+++ pkg/st/deps.go
@@ -9,11 +9,14 @@ import (
 	"runtime"
 	"strings"
 	"sync"
+
+	"charm.land/lipgloss/v2"
+	"github.com/yaklabco/stave/pkg/ui"
 )
 
 // simpleConsoleLogger is an unstructured logger designed for emitting simple
 // messages to the console in `-v`/`--verbose` mode.
-var simpleConsoleLogger = log.New(os.Stderr, "[STAVE] ", 0) //nolint:gochecknoglobals // This is unchanged in the course of the process lifecycle.
+var simpleConsoleLogger = log.New(os.Stderr, lipgloss.NewStyle().Foreground(ui.GetFangScheme().Flag).Render("[STAVE] "), 0) //nolint:gochecknoglobals,lll // This is unchanged in the course of the process lifecycle.
 
 type onceMap struct {
 	mu *sync.Mutex